refactor: list components #3 && tidy up components
