+++ src/components/lists/input-point-item.vue
