+++ src/views/ManualInput.vue
@@ -1,17 +1,13 @@
 <script setup>
 import { onMounted } from 'vue'
-import { useStore } from 'vuex'
 
 import { initializeMap } from '@/utils/map.js'
 
 import GoBack from '@/components/go-back.vue'
-import OptimizeRouteButton from '@/components/optimize-route-button.vue'
+import OptimizeRouteButton from '@/components/buttons/optimize-route-button.vue'
 import CapacityInput from '@/components/capacity-input.vue'
 import PointChooser from '@/components/point-chooser.vue'
-import ListContainer from '@/components/list-container.vue'
-import InputPointItem from '@/components/input-point-item.vue'
-
-const store = useStore()
+import InputPointList from '@/components/lists/input-point-list.vue'
 
 onMounted(() => {
   map.value = initializeMap('map')
@@ -30,11 +26,7 @@ onMounted(() => {
       <div id="map" />
     </div>
 
-    <list-container title="Added input points">
-      <input-point-item v-for="point in store.state.inputPoints"
-                        :key="point.id"
-                        :point="point" />
-    </list-container>
+    <input-point-list />
 
     <div>
       <h1 class="text-center text-2xl font-bold my-4">Input vehicles capacity</h1>
