+++ src/views/ScanQr.vue
@@ -1,14 +1,10 @@
 <script setup>
-import { useStore } from 'vuex'
-
 import GoBack from '@/components/go-back.vue'
 import CodeScanner from '@/components/code-scanner.vue'
-import OptimizeRouteButton from '@/components/optimize-route-button.vue'
+import OptimizeRouteButton from '@/components/buttons/optimize-route-button.vue'
 import CapacityInput from '@/components/capacity-input.vue'
-import ListContainer from '@/components/list-container.vue'
-import InputPointItem from '@/components/input-point-item.vue'
+import InputPointList from '@/components/lists/input-point-list.vue'
 
-const store = useStore()
 </script>
 
 <template>
@@ -18,11 +14,7 @@ const store = useStore()
     <h1 class="text-center text-2xl font-bold my-4">Scan QR Code</h1>
     <code-scanner />
 
-    <list-container title="Added input points">
-      <input-point-item v-for="point in store.state.inputPoints"
-                        :key="point.id"
-                        :point="point" />
-    </list-container>
+   <input-point-list />
 
     <div>
       <h1 class="text-center text-2xl font-bold my-4">Input vehicles capacity</h1>
